finish 4.21

+++ part4/blog-list-backend/controllers/blogs.js
@@ -49,8 +49,23 @@ blogsRouter.post('/', async (request, response) => {
 
 // Delete a blog based on id
 blogsRouter.delete('/:id', async (request, response) => {
-  const existId = await Blog.findByIdAndRemove(request.params.id);
-  existId ? response.status(204).end() : response.status(404).end();
+  const blog = await Blog.findById(request.params.id);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+
+  const user = await User.findById(decodedToken.id);
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    return response.status(401).json({
+      error: 'No permission for deleting the current blog',
+    });
+  }
 });
 
 // Update a blog based on id
